requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
requests
PyMuPDF
